refactor: prefix QRCodeScan component name
