+++ src/frontend/src/lib/components/send/SendQRCodeScan.svelte
